+++ stockrobot/test/utils/PairTest.java
@@ -62,8 +62,8 @@ public class PairTest {
 
 		//Creating some dummy data
 		StockNames stName = new StockNames( "ABB", "LargeCap", true );
-		StockPrices prices = new StockPrices( stName, 3434343, (long)24242492, (long)244424242, (long)224229292, new Date(10101001));
-		StockPrices pricesAnotherInstance = new StockPrices( stName, 3434343, (long)24242492, (long)244424242, (long)224229292, new Date(10101001)); //Same content as above
+		StockPrices prices = new StockPrices( stName, 3434343, 24242492, 244424242, 224229292, new Date(10101001));
+		StockPrices pricesAnotherInstance = new StockPrices( stName, 3434343, 24242492, 244424242, 224229292, new Date(10101001)); //Same content as above
 		
 		//Creating a little tree structure
 		Pair<Pair<String,Pair<String,Integer>>,String> treeStructure = new Pair<Pair<String, Pair<String, Integer>>, String>(
